Move contact requests to crud
