+++ app/controllers/public_controller.rb
@@ -1,6 +1,5 @@
 class PublicController < ApplicationController
-  def index
-  end
+  def index; end
 
   def cookie_policy
     @title = t('public.cookie_policy.title')
@@ -22,38 +21,7 @@ class PublicController < ApplicationController
     @title = t('public.privacy_policy.title')
   end
 
-  def contact_us
-    if request.post?
-      @contact_request = ContactRequest.new(contact_us_params)
-      if @contact_request.save
-        flash[:notice] = t('public.contact_us.flash.contact_request_saved')
-        send_email_to_admins_about_new_request
-      else
-        render 'contact_us'
-      end
-    else
-      @contact_request = ContactRequest.new
-    end
-    @title = t('public.contact_us.title')
-  end
-
   def not_found
     render :not_found, status: :not_found, layout: 'error_page'
-  end
-
-  private
-  def contact_us_params
-    params.require(:contact_request)
-          .permit(
-            :name,
-            :email_address,
-            :telephone,
-            :comments
-          )
-  end
-
-  def send_email_to_admins_about_new_request
-    admins = User.where(is_admin: true)
-    ContactRequestMailer.new_contact_request_notice(admins).deliver_now
   end
 end
