+++ app/views/contact_requests/new.en.html.erb
@@ -3,7 +3,7 @@
 <div class="row">
   <div class="col-md-6">
     <h4>By Email or Chat:</h4>
-    <%= render partial: 'form_contact_us' %>
+    <%= render 'contact_requests/form' %>
   </div>
 
   <div class="col-md-6">
@@ -14,6 +14,6 @@
       <%= OUR_CITY_STATE_ZIP %>
     </p>
 
-    <p>Please use the chat tool in the lower right corner for all other communications.</p>
+    <p class="font-weight-bold">Please use the chat tool in the lower right corner for all other communications.</p>
   </div>
 </div>
