+++ app/views/contact_requests/new.pl.html.erb
@@ -1,2 +1,2 @@
 <h1><%= t('public.contact_us.title') %></h1>
-<%= render 'form_contact_us' %>
+<%= render 'contact_requests/form' %>
