+++ app/views/layouts/_footer.html.erb
@@ -9,9 +9,6 @@
 
     <ul class="nav">
       <li class="nav-item">
-        <%# link_to 'Contact Us', contact_us_path, class: 'nav-link' %>
-      </li>
-      <li class="nav-item">
         <%= link_to t('footer.nav_links.faq'), faq_path, class: 'nav-link' %>
       </li>
       <li class="nav-item">
@@ -24,7 +21,7 @@
         <%= link_to t('footer.nav_links.terms_of_use'), terms_of_use_path, class: 'nav-link' %>
       </li>
       <li class="nav-item">
-        <%= link_to t('footer.nav_links.contact_us'), contact_us_path, class: 'nav-link' %>
+        <%= link_to t('footer.nav_links.contact_us'), new_contact_request_path, class: 'nav-link' %>
       </li>
       <% if user_signed_in? && current_user.is_admin? %>
         <li class="nav-item">
