+++ test/controllers/public_controller_test.rb
@@ -2,14 +2,6 @@ require 'test_helper'
 
 class PublicControllerTest < ActionDispatch::IntegrationTest
   setup do
-    @contact_us_params = {
-      contact_request: {
-        name: 'Test Name',
-        email_address: 'test@test.com',
-        telephone: '123456789'
-      }
-    }
-    @admin_one = create(:user, :admin)
   end
 
   test 'should get index' do
@@ -35,72 +27,6 @@ class PublicControllerTest < ActionDispatch::IntegrationTest
   test 'should get terms_of_use' do
     get terms_of_use_url
     assert_response :success
-  end
-
-  test 'should get contact_us' do
-    get contact_us_url
-    assert_response :success
-  end
-
-  test 'it creates contact request with correct params' do
-    assert_difference('ContactRequest.count') do
-      post contact_us_url, params: @contact_us_params
-    end
-    assert_response :success
-  end
-
-  test 'do not create contact request with name longer than 20 characters' do
-    @contact_us_params[:contact_request][:name] = \
-      Faker::Internet.username(specifier: 21)
-
-    assert_no_difference('ContactRequest.count') do
-      post contact_us_url, params: @contact_us_params
-      assert_response :success
-    end
-  end
-
-  test 'it does not create contact request with empty email' do
-    @contact_us_params[:contact_request][:email_address] = ''
-    assert_no_difference('ContactRequest.count') do
-      post contact_us_url, params: @contact_us_params
-      assert_response :success
-    end
-  end
-
-  test 'do not create contact request with email longer than 50 characters' do
-    @contact_us_params[:contact_request][:email_address] = \
-      Faker::Internet.email(
-        name: Faker::Internet.username(specifier: 50),
-        separators: '_'
-      )
-    assert_no_difference('ContactRequest.count') do
-      post contact_us_url, params: @contact_us_params
-      assert_response :success
-    end
-  end
-
-  test 'it does not create contact request with email in wrong format' do
-    @contact_us_params[:contact_request][:email_address] = \
-      'wrong_email_format.com'
-    assert_no_difference('ContactRequest.count') do
-      post contact_us_url, params: @contact_us_params
-      assert_response :success
-    end
-  end
-
-  test 'do not create contact request with comments longer than 500 chars' do
-    @contact_us_params[:contact_request][:comments] = \
-      Faker::Lorem.characters(number: 501)
-    assert_no_difference('ContactRequest.count') do
-      post contact_us_url, params: @contact_us_params
-      assert_response :success
-    end
-  end
-
-  test 'should send email to admins' do
-    assert_difference('ActionMailer::Base.deliveries.count') do
-      post contact_us_url, params: @contact_us_params
-    end
   end
 
   test "wrong path renders not_found" do
