--- test/fixtures/contact_requests.yml
@@ -1,1 +1,0 @@
-# Read about fixtures at https://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
